move proposal focalboard files into folder
