+++ lib/focalboard/proposals/__tests__/extractCardProposalProperties.spec.ts
