+++ lib/focalboard/proposals/__tests__/extractDatabaseProposalProperties.spec.ts
