+++ lib/focalboard/proposals/__tests__/setDatabaseProposalProperties.spec.ts
