+++ lib/focalboard/proposals/extractCardProposalProperties.ts
