+++ lib/focalboard/proposals/extractDatabaseProposalProperties.ts
