+++ lib/focalboard/proposals/proposalDbProperties.ts
